requirements.txt
numpy
tensorflow
pillow
fastapi
firebase_admin
pymongo
pika